[#124] refactor: Card 요청에 대한 API 분리
Card 요청에 대한 API를 분리하였습니다.
